+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/_params.py
@@ -213,6 +213,11 @@ def load_arguments(self, _):
         c.argument('os_type', get_enum_type(aci_connector_os_type),
                    help='The OS type of the connector')
 
+    with self.argument_context('aks update-credentials') as c:
+        c.argument('reset_service_principal', action='store_true')
+        c.argument('service_principal')
+        c.argument('client_secret')
+
     with self.argument_context('aks upgrade') as c:
         c.argument('kubernetes_version', completer=get_k8s_upgrades_completion_list)