+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -9,6 +9,7 @@ Release History
 * Add Managed OpenShift commands
 * Minor fixes
 * Add support for new ACI regions
+* Support Service Principal updating operation using "az aks update-credentials --reset-service-principal"
 
 2.3.13
 ++++++